Move tests for geodetic representation to separate file
